Move text layout logic out of Document
